backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
azure-core==1.32.0
azure-storage-blob==12.24.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
distro==1.9.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.9.0
openai==1.66.3
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyodbc==5.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
